Rename the "document_id" field to "d_id".
For applications with large numbers of n-grams, this will significantly
reduce the amount of storage and memory used by mongoid_fulltext.

For small numbers of n-grams, this will have no noticable effect.

I usually prefer explicitness, but this is necessary for pure
performance reasons when dealing with millions of n-grams.